♻️ Migrate routes to React Router v7 file-based structure
