+++ app/routes/api.thumbnail-encrypted.$id.ts
@@ -1,0 +1,1 @@
+export { loader } from '~/modules/thumbnail/decrypt-thumbnail/decrypt-thumbnail.route';
