+++ app/routes/api.thumbnail-preview.$filename.ts
