+++ app/routes/api.thumbnail.$id.ts
